Move list.cpp's prototypes into their own header
